Add AppletAE, step 1: move common interfaces to am.h
