+++ src/core/hle/service/am/am.h
@@ -5,7 +5,12 @@
 #pragma once
 
 #include <memory>
+#include "core/hle/ipc_helpers.h"
 #include "core/hle/service/service.h"
+
+namespace Kernel {
+class Event;
+}
 
 namespace Service {
 namespace NVFlinger {
@@ -13,6 +18,95 @@ class NVFlinger;
 }
 
 namespace AM {
+
+// TODO: Add more languages
+enum SystemLanguage {
+    Japanese = 0,
+    English = 1,
+};
+
+class IWindowController final : public ServiceFramework<IWindowController> {
+public:
+    IWindowController();
+
+private:
+    void GetAppletResourceUserId(Kernel::HLERequestContext& ctx);
+    void AcquireForegroundRights(Kernel::HLERequestContext& ctx);
+};
+
+class IAudioController final : public ServiceFramework<IAudioController> {
+public:
+    IAudioController();
+};
+
+class IDisplayController final : public ServiceFramework<IDisplayController> {
+public:
+    IDisplayController();
+};
+
+class IDebugFunctions final : public ServiceFramework<IDebugFunctions> {
+public:
+    IDebugFunctions();
+};
+
+class ISelfController final : public ServiceFramework<ISelfController> {
+public:
+    ISelfController(std::shared_ptr<NVFlinger::NVFlinger> nvflinger);
+
+private:
+    void SetFocusHandlingMode(Kernel::HLERequestContext& ctx);
+    void SetRestartMessageEnabled(Kernel::HLERequestContext& ctx);
+    void SetPerformanceModeChangedNotification(Kernel::HLERequestContext& ctx);
+    void SetOperationModeChangedNotification(Kernel::HLERequestContext& ctx);
+    void SetOutOfFocusSuspendingEnabled(Kernel::HLERequestContext& ctx);
+    void LockExit(Kernel::HLERequestContext& ctx);
+    void UnlockExit(Kernel::HLERequestContext& ctx);
+    void CreateManagedDisplayLayer(Kernel::HLERequestContext& ctx);
+
+    std::shared_ptr<NVFlinger::NVFlinger> nvflinger;
+};
+
+class ICommonStateGetter final : public ServiceFramework<ICommonStateGetter> {
+public:
+    ICommonStateGetter();
+
+private:
+    enum class FocusState : u8 {
+        InFocus = 1,
+        NotInFocus = 2,
+    };
+
+    enum class OperationMode : u8 {
+        Handheld = 0,
+        Docked = 1,
+    };
+
+    void GetEventHandle(Kernel::HLERequestContext& ctx);
+    void ReceiveMessage(Kernel::HLERequestContext& ctx);
+    void GetCurrentFocusState(Kernel::HLERequestContext& ctx);
+    void GetOperationMode(Kernel::HLERequestContext& ctx);
+    void GetPerformanceMode(Kernel::HLERequestContext& ctx);
+
+    Kernel::SharedPtr<Kernel::Event> event;
+};
+
+class ILibraryAppletCreator final : public ServiceFramework<ILibraryAppletCreator> {
+public:
+    ILibraryAppletCreator();
+};
+
+class IApplicationFunctions final : public ServiceFramework<IApplicationFunctions> {
+public:
+    IApplicationFunctions();
+
+private:
+    void PopLaunchParameter(Kernel::HLERequestContext& ctx);
+    void SetTerminateResult(Kernel::HLERequestContext& ctx);
+    void GetDesiredLanguage(Kernel::HLERequestContext& ctx);
+    void InitializeGamePlayRecording(Kernel::HLERequestContext& ctx);
+    void SetGamePlayRecordingState(Kernel::HLERequestContext& ctx);
+    void NotifyRunning(Kernel::HLERequestContext& ctx);
+};
 
 /// Registers all AM services with the specified service manager.
 void InstallInterfaces(SM::ServiceManager& service_manager,
