+++ src/core/hle/service/am/applet_oe.h
@@ -15,12 +15,6 @@ class NVFlinger;
 
 namespace AM {
 
-// TODO: Add more languages
-enum SystemLanguage {
-    Japanese = 0,
-    English = 1,
-};
-
 class AppletOE final : public ServiceFramework<AppletOE> {
 public:
     AppletOE(std::shared_ptr<NVFlinger::NVFlinger> nvflinger);
